+++ packages/adapter-mariadb/src/mariadb.test.ts
@@ -1,5 +1,5 @@
 import * as mariadb from 'mariadb'
-import { describe, expect, test, vi, beforeAll, afterAll } from 'vitest'
+import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest'
 
 import {
   inferCapabilities,
@@ -7,73 +7,6 @@ import {
   PrismaMariaDbAdapterFactory,
   rewriteConnectionString,
 } from './mariadb'
-
-describe.each([
-  ['8.0.12', { supportsRelationJoins: false }],
-  ['8.0.13', { supportsRelationJoins: true }],
-  ['8.1.0', { supportsRelationJoins: true }],
-  ['8.4.5', { supportsRelationJoins: true }],
-  ['8.4.13', { supportsRelationJoins: true }],
-  ['11.4.7-MariaDB-ubu2404', { supportsRelationJoins: false }],
-])('infer capabilities for %s', (version, capabilities) => {
-  test(`inferCapabilities(${version})`, () => {
-    expect(inferCapabilities(version)).toEqual(capabilities)
-  })
-})
-
-describe('rewriteConnectionString', () => {
-  test('should rewrite mysql:// to mariadb://', () => {
-    const input = 'mysql://user:password@localhost:3306/database?ssl=true&connectionLimit=10&charset=utf8mb4'
-    const expected = 'mariadb://user:password@localhost:3306/database?ssl=true&connectionLimit=10&charset=utf8mb4'
-    expect(rewriteConnectionString(new URL(input)).toString()).toBe(expected)
-  })
-
-  test('should preserve mariadb:// connection strings', () => {
-    const input = 'mariadb://user:pass@localhost:3306/db'
-    expect(rewriteConnectionString(new URL(input)).toString()).toBe(input)
-  })
-})
-
-describe('credential sanitization', () => {
-  test('connection string parse error should not expose password', async () => {
-    const secretPassword = 'super_secret_password_12345'
-    // IPv6 address in brackets - causes parse error in mariadb driver
-    const connectionString = `mariadb://user:${secretPassword}@[64:ff9b::23be:d64c]/db`
-
-    const factory = new PrismaMariaDbAdapterFactory(connectionString)
-
-    try {
-      await factory.connect()
-      expect.fail('Expected connection to fail')
-    } catch (error) {
-      const errorMessage = String(error)
-      expect(errorMessage).not.toContain(secretPassword)
-    }
-  })
-})
-
-describe('useTextProtocol option', () => {
-  const flagToMethod = {
-    true: 'query',
-    false: 'execute',
-  }
-
-  test.each([false, undefined, true].map((flag) => ({ flag, method: flagToMethod[String(!!flag)] })))(
-    'should use client.$method when useTextProtocol is $flag',
-    async ({ flag, method }) => {
-      const mockClient = {
-        execute: vi.fn().mockResolvedValue({ meta: [], affectedRows: 1 }),
-        query: vi.fn().mockResolvedValue({ meta: [], affectedRows: 1 }),
-      } as unknown as mariadb.Pool
-
-      const adapter = new PrismaMariaDbAdapter(mockClient, { supportsRelationJoins: true }, { useTextProtocol: flag })
-      await adapter.executeRaw({ sql: 'SELECT 1', args: [], argTypes: [] })
-      expect(mockClient[method]).toHaveBeenCalledWith(expect.objectContaining({ sql: 'SELECT 1' }), [])
-
-      expect(mockClient[flagToMethod[String(!flag)]]).not.toHaveBeenCalled()
-    },
-  )
-})
 
 describe('PrismaMariaDbAdapterFactory constructor', () => {
   beforeAll(() => {
@@ -122,4 +55,53 @@ describe('PrismaMariaDbAdapterFactory constructor', () => {
     expect(mockCreatePool).toHaveBeenCalledWith(expected)
     mockCreatePool.mockClear()
   })
+})
+
+describe.each([
+  ['8.0.12', { supportsRelationJoins: false }],
+  ['8.0.13', { supportsRelationJoins: true }],
+  ['8.1.0', { supportsRelationJoins: true }],
+  ['8.4.5', { supportsRelationJoins: true }],
+  ['8.4.13', { supportsRelationJoins: true }],
+  ['11.4.7-MariaDB-ubu2404', { supportsRelationJoins: false }],
+])('infer capabilities for %s', (version, capabilities) => {
+  test(`inferCapabilities(${version})`, () => {
+    expect(inferCapabilities(version)).toEqual(capabilities)
+  })
+})
+
+describe('rewriteConnectionString', () => {
+  test('should rewrite mysql:// to mariadb://', () => {
+    const input = 'mysql://user:password@localhost:3306/database?ssl=true&connectionLimit=10&charset=utf8mb4'
+    const expected = 'mariadb://user:password@localhost:3306/database?ssl=true&connectionLimit=10&charset=utf8mb4'
+    expect(rewriteConnectionString(new URL(input)).toString()).toBe(expected)
+  })
+
+  test('should preserve mariadb:// connection strings', () => {
+    const input = 'mariadb://user:pass@localhost:3306/db'
+    expect(rewriteConnectionString(new URL(input)).toString()).toBe(input)
+  })
+})
+
+describe('useTextProtocol option', () => {
+  const flagToMethod = {
+    true: 'query',
+    false: 'execute',
+  }
+
+  test.each([false, undefined, true].map((flag) => ({ flag, method: flagToMethod[String(!!flag)] })))(
+    'should use client.$method when useTextProtocol is $flag',
+    async ({ flag, method }) => {
+      const mockClient = {
+        execute: vi.fn().mockResolvedValue({ meta: [], affectedRows: 1 }),
+        query: vi.fn().mockResolvedValue({ meta: [], affectedRows: 1 }),
+      } as unknown as mariadb.Pool
+
+      const adapter = new PrismaMariaDbAdapter(mockClient, { supportsRelationJoins: true }, { useTextProtocol: flag })
+      await adapter.executeRaw({ sql: 'SELECT 1', args: [], argTypes: [] })
+      expect(mockClient[method]).toHaveBeenCalledWith(expect.objectContaining({ sql: 'SELECT 1' }), [])
+
+      expect(mockClient[flagToMethod[String(!flag)]]).not.toHaveBeenCalled()
+    },
+  )
 })
